+++ tests/test_geometry.py
@@ -1,5 +1,6 @@
 import pytest
 
+from pycaption import CaptionReadSyntaxError
 from pycaption.geometry import Size, Point, Stretch, Padding, UnitEnum, Layout
 
 
@@ -121,3 +122,22 @@ class TestIsRelative:
         assert not layout_abs.is_relative()
         assert not layout_mix.is_relative()
         assert layout_rel.is_relative()
+
+
+class TestSize:
+    @pytest.mark.parametrize('string, value, unit', [
+        ('1px', 1.0, UnitEnum.PIXEL), ('2.3em', 2.3, UnitEnum.EM),
+        ('12.34%', 12.34, UnitEnum.PERCENT), ('1.234c', 1.234, UnitEnum.CELL),
+        ('10pt', 10.0, UnitEnum.PT), ('0', 0.0, UnitEnum.PIXEL)])
+    def test_valid_size_from_string(self, string, value, unit):
+        size = Size.from_string(string)
+
+        assert size.value == value
+        assert size.unit == unit
+
+    @pytest.mark.parametrize('string', ['10', '11,1px', '12xx', '%', 'o1pt'])
+    def test_invalid_size_from_string(self, string):
+        with pytest.raises(CaptionReadSyntaxError) as exc_info:
+            Size.from_string(string)
+
+        assert exc_info.value.args[0].startswith(f"Invalid size: {string}.")